requirements.txt
atomic-agents>=1.0.0,<2.0.0
pydantic>=2.8.2,<3.0.0
sympy>=1.12,<2.0.0
aiohttp>=3.9.0,<4.0.0
beautifulsoup4>=4.12.0,<5.0.0
markdownify>=0.11.0,<1.0.0
readability-lxml>=0.8.1,<1.0.0
requests>=2.31.0,<3.0.0
lxml>=5.1.0,<6.0.0
lxml-html-clean>=0.3.1,<1.0.0
instructor>=1.7.0
openai>=1.5.0
python-dotenv>=1.0.0
rich>=13.0.0
email_validator>=2.0.0

firecrawl-py
matplotlib
numpy
pandas
json5~=0.10.0
tenacity~=9.0.0

